refact(mysql): init mysql package
