+++ cmd/matching.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"dcard/service/matching"
+	"dcard/storage/mysql"
 
 	"github.com/spf13/cobra"
 )
@@ -18,6 +19,9 @@ func init() {
 }
 
 func RunMatchingCmd(cmd *cobra.Command, args []string) error {
+	mysql.Initialize()
+
 	matching.EnableMatchingServer()
+
 	return nil
 }
