+++ service/post/api_memberDataGet.go
@@ -1,16 +1,16 @@
 package post
 
 import (
+	"dcard/storage/mysql"
 	"encoding/json"
 	"net/http"
 )
 
 func edit(w http.ResponseWriter, r *http.Request) {
 	// Authentication(w, r)
-	db := mysqlConn()
 	creds := &Member{}
 	err := json.NewDecoder(r.Body).Decode(creds)
-	selDB, err := db.Query("SELECT * FROM Member WHERE MemberID=? LIMIT 1", creds.MemberID)
+	selDB, err := mysql.GetMySQL().Query("SELECT * FROM Member WHERE MemberID=? LIMIT 1", creds.MemberID)
 	mem := Member{}
 	for selDB.Next() {
 		var MemberID int
@@ -42,7 +42,6 @@ func edit(w http.ResponseWriter, r *http.Request) {
 		mem.CreateDate = CreateDate
 	}
 	// tmpl.ExecuteTemplate(w, "Edit", mem)
-	defer db.Close()
 	a, err := json.Marshal(mem)
 	w.Write(a)
 	return
